# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DocManageApp API",
    "version": "1.2.0",
    "description": "💼 **Agrobank Executive Discipline System API** — единая платформа для управления документами, переписками, пользователями и ролями внутри организации.",
    "contact": {
      "name": "Agrobank IT Department",
      "email": "support@agrobank.uz",
      "url": "https://agrobank.uz"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    { "url": "https://api.agrobank.uz/v1", "description": "Production Server" },
    { "url": "https://staging.agrobank.uz/v1", "description": "Staging Server" }
  ],
...
  "paths": {
...
    "/correspondences": {
      "get": {
        "summary": "Получить корреспонденции текущего пользователя",
        "description": "Результат зависит от роли авторизованного пользователя. Менеджеры видят входящие, сотрудники — назначенные им документы.",
        "operationId": "getCorrespondences",
        "tags": ["Correspondence"],
        "responses": {
          "200": {
            "description": "Список корреспонденций",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Correspondence" } },
                "examples": {
                  "incoming": {
                    "summary": "Пример входящей корреспонденции",
                    "value": [
                      {
                        "id": 1,
                        "type": "Kiruvchi",
                        "title": "Tahlil hisobotlari",
                        "content": "Departamentdan kelgan hujjatlar.",
                        "source": "HR Department",
                        "stage": "Yangi",
                        "deadline": "2025-11-10T00:00:00Z",
                        "department": "Bosh ofis",
                        "mainExecutorId": 2
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Correspondence": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "type": { "type": "string", "enum": ["Kiruvchi", "Chiquvchi"], "example": "Kiruvchi" },
          "title": { "type": "string", "example": "Hisobot" },
          "content": { "type": "string", "example": "2025-yil 1-chorak faoliyati bo‘yicha hisobot" },
          "source": { "type": "string", "example": "Markaziy bo‘lim" },
          "stage": { "type": "string", "example": "Yangi" },
          "deadline": { "type": "string", "format": "date-time", "example": "2025-11-10T00:00:00Z" },
          "department": { "type": "string", "example": "Bosh ofis" },
          "mainExecutorId": { "type": "integer", "example": 2 }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": { "type": "integer", "example": 400 },
          "message": { "type": "string", "example": "Invalid input data" },
          "details": { "type": "string", "example": "Email format is invalid" }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Неавторизованный доступ",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }]
}